api/requirements.txt
fastapi
uvicorn
boto3
python-dotenv
redis
PyMuPDF
Pillow
requests
python-multipart
